Added SerialSetCtrlRequest
Allow the RTS and DTR to set over the WebUI.
DIY Tracker fix for some D1 mini with CH340 chipsets that use different control pins.

Adding SerialRestartTrackerRequest

This is to Invoke a Try to Restart the Tracker over the Serial Port via RTS / DTR.
This helps to see the full log in the WiFi window.
This does not work with the SlimeVR Tracker